Refactor table view impl by reusing reader

+++ pulsar-client-api/src/main/java/org/apache/pulsar/client/api/ReaderBuilder.java
@@ -311,7 +311,7 @@ public interface ReaderBuilder<T> extends Cloneable {
      *            the time unit of the interval.
      * @return the reader builder instance
      */
-    ReaderBuilder<T> autoUpdatePartitionsInterval(int interval, TimeUnit unit);
+    ReaderBuilder<T> autoUpdatePartitionsInterval(long interval, TimeUnit unit);
 
     /**
      * Intercept {@link Reader}.
